Close deflater and inflater streams on connection close
Without context takeover, the deflater and inflater are initialized and closed
on every message. However, with context takeover, the deflater and inflater
persist for the entire lifetime of the connection. They need to be closed on a
connection close.

The deflater is on a ChannelOutboundHandler and must be closed when the
WebSocketConnections issues a channel.close(). The inflater is on a
ChannelInboundHandler, it must be closed when we receive a `closeConnection`
frame from the remote peer.